the source code
this code use google mobilenet for object detect, and then get the english words about this object. The accuracy is not that good, may be just for fun. The build file in the app file is a big size, then i don't want upload, and you may not need it. The mobilenet pb you can download from google.